@@ -1,6 +1,7 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
+import 'package:musico/modals/playlist_category.dart';
 import 'package:musico/screens/playlist_screen.dart';
 import 'package:page_transition/page_transition.dart';
 
@@ -9,10 +10,12 @@ class SongCard extends StatefulWidget {
   String playlistname;
   String thumbnail;
   String playlistid;
+  Description playlistdesc;
 
   SongCard(
       {super.key,
       required this.playlistname,
+      required this.playlistdesc,
       required this.thumbnail,
       required this.playlistid});
 
@@ -28,42 +31,66 @@ class _SongCardState extends State<SongCard> {
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      children: [
-        ClipRRect(
-          borderRadius: const BorderRadius.all(Radius.circular(10.0)),
-          child: InkWell(
-            child: Container(
-              width: 160,
-              height: 160,
-              decoration: BoxDecoration(
-                image: DecorationImage(
-                  fit: BoxFit.cover,
-                  image: CachedNetworkImageProvider(widget.thumbnail),
+    return Container(
+      decoration: BoxDecoration(
+          color: const Color(0XFF1e1c22),
+          borderRadius: BorderRadius.circular(20.0)),
+      child: Column(
+        children: [
+          ClipRRect(
+            borderRadius: const BorderRadius.all(Radius.circular(20.0)),
+            child: InkWell(
+              child: Container(
+                width: 160,
+                height: 160,
+                decoration: BoxDecoration(
+                  image: DecorationImage(
+                    fit: BoxFit.cover,
+                    image: CachedNetworkImageProvider(widget.thumbnail),
+                  ),
                 ),
               ),
+              onTap: () {
+                Navigator.push(
+                    context,
+                    PageTransition(
+                        duration: const Duration(milliseconds: 400),
+                        child: PlaylistScreen(playlistId: widget.playlistid),
+                        type: PageTransitionType.rightToLeftWithFade));
+              },
             ),
-            onTap: () {
-              Navigator.push(
-                  context,
-                  PageTransition(
-                      duration: const Duration(milliseconds: 400),
-                      child: PlaylistScreen(playlistId: widget.playlistid),
-                      type: PageTransitionType.rightToLeftWithFade));
-            },
           ),
-        ),
-        SizedBox(
-          width: 150,
-          child: Text(
-            widget.playlistname,
-            maxLines: 1,
-            overflow: TextOverflow.ellipsis,
-            style: GoogleFonts.nunitoSans(
-                color: Colors.white, fontSize: 16, fontWeight: FontWeight.w700),
+          const Spacer(),
+          SizedBox(
+            width: 160,
+            child: Text(
+              widget.playlistname,
+              maxLines: 1,
+              textAlign: TextAlign.center,
+              overflow: TextOverflow.ellipsis,
+              style: GoogleFonts.nunitoSans(
+                  color: Colors.white,
+                  fontSize: 16,
+                  fontWeight: FontWeight.w700),
+            ),
           ),
-        ),
-      ],
+          const Spacer(),
+          SizedBox(
+            width: 160,
+            child: Text(
+              "Playlist • YouTube Music",
+              maxLines: 1,
+              textAlign: TextAlign.center,
+              overflow: TextOverflow.ellipsis,
+              style: GoogleFonts.nunitoSans(
+                color: Colors.white60,
+                fontSize: 12,
+              ),
+            ),
+          ),
+          const Spacer()
+        ],
+      ),
     );
 
     /*
