@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:just_audio_background/just_audio_background.dart';
-import 'package:musico/screens/home_screen.dart';
+import 'package:musico/screens/base_screen.dart';
 
 Future<void> main() async {
   await JustAudioBackground.init(
@@ -27,12 +27,13 @@ class _MyAppState extends State<MyApp> {
       title: 'Musico',
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
+        primarySwatch: Colors.deepPurple,
         textTheme: Theme.of(context).textTheme.apply(
               bodyColor: Colors.white,
               displayColor: Colors.white,
             ),
       ),
-      home: const HomeScreen(),
+      home: const BaseScreen(),
     );
   }
 }
