@@ -1,5 +1,5 @@
 import 'dart:convert';
-import 'package:get/get.dart';
+import 'package:bootstrap_icons/bootstrap_icons.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:http/http.dart' as http;
@@ -28,137 +28,264 @@ class _SearchResultScreenState extends State<SearchResultScreen> {
                 begin: Alignment.topRight,
                 end: Alignment.topLeft,
                 colors: [
-              Color(0XFF780206),
-              Color(0XFF061161),
+              // Color(0XFF780206),
+              // Color(0XFF061161),
+              Colors.black,
+              Colors.black
             ])),
-        child: Scaffold(
-          appBar: AppBar(
-            backgroundColor: Colors.black38,
-            elevation: 0,
-            leading: IconButton(
-              icon: const Icon(
-                Icons.arrow_back_ios,
-              ),
-              onPressed: () {
-                Get.back();
-              },
-            ),
-            actions: [
-              Padding(
-                padding: const EdgeInsets.only(right: 15.0),
-                child: Text(
-                  'Songs',
-                  style: GoogleFonts.poppins(
-                    fontSize: 35,
-                    fontWeight: FontWeight.bold,
+        child: DefaultTabController(
+          length: 2,
+          child: Scaffold(
+              appBar: AppBar(
+                backgroundColor: Colors.black26,
+                elevation: 0,
+                bottom: const TabBar(
+                    indicatorColor: Color(0XFFC4FC4C),
+                    labelColor: Color(0XFFC4FC4C),
+                    unselectedLabelColor: Colors.grey,
+                    tabs: [
+                      Tab(
+                          icon: Icon(
+                        BootstrapIcons.music_note,
+                      )),
+                      Tab(
+                        child: Icon(
+                          Icons.movie,
+                        ),
+                      )
+                    ]),
+                leading: IconButton(
+                  onPressed: () {
+                    Navigator.pop(context);
+                  },
+                  icon: const Icon(Icons.arrow_back_ios_new),
+                ),
+                actions: [
+                  Padding(
+                    padding: const EdgeInsets.only(right: 15.0),
+                    child: Text(
+                      'Songs',
+                      style: GoogleFonts.poppins(
+                        fontSize: 35,
+                      ),
+                      textAlign: TextAlign.center,
+                    ),
                   ),
-                  textAlign: TextAlign.center,
+                ],
+              ),
+              backgroundColor: Colors.black26,
+              floatingActionButton: FloatingActionButton(
+                onPressed: () {
+                  Navigator.pop(context);
+                },
+                backgroundColor: const Color(0XFFC4FC4C),
+                child: const Icon(
+                  Icons.search,
+                  color: Colors.black,
                 ),
               ),
-            ],
-          ),
-          backgroundColor: Colors.black26,
-          floatingActionButton: FloatingActionButton(
-            onPressed: () {
-              Navigator.pop(context);
-            },
-            backgroundColor: const Color.fromARGB(255, 0, 0, 0),
-            child: const Icon(
-              Icons.search,
-              color: Colors.white,
-            ),
-          ),
-          body: FutureBuilder(
-            future: getData(widget.message),
-            builder: (context, snapshot) {
-              if (snapshot.connectionState == ConnectionState.waiting) {
-                return ListView.builder(
-                  itemCount: 15, // Number of shimmer placeholders
-                  itemBuilder: (context, index) {
-                    return const SearchResultShimmer(); // Shimmer placeholder tile
-                  },
-                );
-              } else if (snapshot.hasData) {
-                return ListView.separated(
-                  separatorBuilder: (context, index) {
-                    return const SizedBox(
-                        width: 10); // Add a horizontal spacing between items
-                  },
-                  itemCount: snapshot.data!.length,
-                  itemBuilder: (context, index) {
-                    return ListTile(
-                      leading: ClipRRect(
-                        borderRadius: BorderRadius.circular(5.0),
-                        child: Image.network(
-                          snapshot.data![index]["thumbnails"].contains("?sqp=")
-                              ? snapshot.data![index]["thumbnails"]
-                                  .split("?sqp=")[0]
-                              : snapshot.data![index]["thumbnails"],
-                          width: 100,
-                          height: 100,
-                          fit: BoxFit.cover,
-                        ),
-                      ),
-                      title: Text(
-                        snapshot.data![index]["title"],
-                        style: GoogleFonts.nunito(
-                          color: Colors.white,
-                          fontWeight: FontWeight.bold,
-                        ),
-                        overflow: TextOverflow.ellipsis,
-                      ),
-                      subtitle: Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        children: [
-                          Expanded(
-                            child: Text(
-                              snapshot.data![index]["artists"],
-                              style: GoogleFonts.poppins(color: Colors.white70),
-                              overflow: TextOverflow.ellipsis,
-                            ),
+              body: TabBarView(
+                children: [
+                  FutureBuilder(
+                    future: getData(widget.message, "songs"),
+                    builder: (context, snapshot) {
+                      if (snapshot.connectionState == ConnectionState.waiting) {
+                        return ListView.builder(
+                          itemCount: 15, // Number of shimmer placeholders
+                          itemBuilder: (context, index) {
+                            return const SearchResultShimmer(); // Shimmer placeholder tile
+                          },
+                        );
+                      } else if (snapshot.hasData) {
+                        return ListView.separated(
+                          separatorBuilder: (context, index) {
+                            return const SizedBox(
+                                width:
+                                    10); // Add a horizontal spacing between items
+                          },
+                          itemCount: snapshot.data!.length,
+                          itemBuilder: (context, index) {
+                            return ListTile(
+                              splashColor: Colors.white38,
+                              enabled: true,
+                              leading: ClipRRect(
+                                borderRadius: BorderRadius.circular(5.0),
+                                child: Image.network(
+                                  snapshot.data![index]["thumbnails"]
+                                          .contains("?sqp=")
+                                      ? snapshot.data![index]["thumbnails"]
+                                          .split("?sqp=")[0]
+                                      : snapshot.data![index]["thumbnails"]
+                                              .contains("=w120")
+                                          ? snapshot.data![index]["thumbnails"]
+                                                  .split("=w120")[0] +
+                                              "=w240-h240-l90-rj"
+                                          : snapshot.data![index]["thumbnails"],
+                                  width: 100,
+                                  height: 100,
+                                  fit: BoxFit.cover,
+                                ),
+                              ),
+                              title: Text(
+                                snapshot.data![index]["title"],
+                                style: GoogleFonts.nunito(
+                                  color: Colors.white,
+                                  fontWeight: FontWeight.bold,
+                                ),
+                                overflow: TextOverflow.ellipsis,
+                              ),
+                              subtitle: Row(
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceBetween,
+                                children: [
+                                  Expanded(
+                                    child: Text(
+                                      snapshot.data![index]["artists"],
+                                      style: GoogleFonts.poppins(
+                                          color: Colors.white70),
+                                      overflow: TextOverflow.ellipsis,
+                                    ),
+                                  ),
+                                  Text(
+                                    snapshot.data![index]["duration"] ?? '0:00',
+                                    style: GoogleFonts.poppins(
+                                        color: Colors.white70),
+                                    overflow: TextOverflow.ellipsis,
+                                  )
+                                ],
+                              ),
+                              tileColor: const Color(0XFF1e1c22),
+                              onTap: () {
+                                Navigator.push(
+                                    context,
+                                    PageTransition(
+                                        child: PlayerScreen(
+                                            vd: snapshot.data![index]
+                                                ["videoId"]),
+                                        type: PageTransitionType.bottomToTop,
+                                        duration:
+                                            const Duration(milliseconds: 300)));
+                              },
+                            );
+                          },
+                        );
+                      } else {
+                        return const Center(
+                          child: Text(
+                            'No data available.',
+                            style: TextStyle(color: Colors.white),
                           ),
-                          Text(
-                            snapshot.data![index]["duration"] ?? '0:00',
-                            style: GoogleFonts.poppins(color: Colors.white70),
-                            overflow: TextOverflow.ellipsis,
-                          )
-                        ],
-                      ),
-                      tileColor: Colors.black12,
-                      onTap: () {
-                        Navigator.push(
-                            context,
-                            PageTransition(
-                                child: PlayerScreen(
-                                    vd: snapshot.data![index]["videoId"]),
-                                type: PageTransitionType.bottomToTop,
-                                duration: const Duration(milliseconds: 300)));
-                      },
-                    );
-                  },
-                );
-              } else {
-                return const Center(
-                  child: Text(
-                    'No data available.',
-                    style: TextStyle(color: Colors.white),
+                        );
+                      }
+                    },
                   ),
-                );
-              }
-            },
-          ),
+                  FutureBuilder(
+                    future: getData(widget.message, "videos"),
+                    builder: (context, snapshot) {
+                      if (snapshot.connectionState == ConnectionState.waiting) {
+                        return ListView.builder(
+                          itemCount: 15, // Number of shimmer placeholders
+                          itemBuilder: (context, index) {
+                            return const SearchResultShimmer(); // Shimmer placeholder tile
+                          },
+                        );
+                      } else if (snapshot.hasData) {
+                        return ListView.separated(
+                          separatorBuilder: (context, index) {
+                            return const SizedBox(
+                                width:
+                                    10); // Add a horizontal spacing between items
+                          },
+                          itemCount: snapshot.data!.length,
+                          itemBuilder: (context, index) {
+                            return ListTile(
+                              splashColor: Colors.white38,
+                              enabled: true,
+                              leading: ClipRRect(
+                                borderRadius: BorderRadius.circular(5.0),
+                                child: Image.network(
+                                  snapshot.data![index]["thumbnails"]
+                                          .contains("?sqp=")
+                                      ? snapshot.data![index]["thumbnails"]
+                                          .split("?sqp=")[0]
+                                      : snapshot.data![index]["thumbnails"]
+                                              .contains("=w120")
+                                          ? snapshot.data![index]["thumbnails"]
+                                                  .split("=w120")[0] +
+                                              "=w240-h240-l90-rj"
+                                          : snapshot.data![index]["thumbnails"],
+                                  width: 100,
+                                  height: 100,
+                                  fit: BoxFit.cover,
+                                ),
+                              ),
+                              title: Text(
+                                snapshot.data![index]["title"],
+                                style: GoogleFonts.nunito(
+                                  color: Colors.white,
+                                  fontWeight: FontWeight.bold,
+                                ),
+                                overflow: TextOverflow.ellipsis,
+                              ),
+                              subtitle: Row(
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceBetween,
+                                children: [
+                                  Expanded(
+                                    child: Text(
+                                      snapshot.data![index]["artists"],
+                                      style: GoogleFonts.poppins(
+                                          color: Colors.white70),
+                                      overflow: TextOverflow.ellipsis,
+                                    ),
+                                  ),
+                                  Text(
+                                    snapshot.data![index]["duration"] ?? '0:00',
+                                    style: GoogleFonts.poppins(
+                                        color: Colors.white70),
+                                    overflow: TextOverflow.ellipsis,
+                                  )
+                                ],
+                              ),
+                              tileColor: const Color(0XFF1e1c22),
+                              onTap: () {
+                                Navigator.push(
+                                    context,
+                                    PageTransition(
+                                        child: PlayerScreen(
+                                            vd: snapshot.data![index]
+                                                ["videoId"]),
+                                        type: PageTransitionType.bottomToTop,
+                                        duration:
+                                            const Duration(milliseconds: 300)));
+                              },
+                            );
+                          },
+                        );
+                      } else {
+                        return const Center(
+                          child: Text(
+                            'No data available.',
+                            style: TextStyle(color: Colors.white),
+                          ),
+                        );
+                      }
+                    },
+                  )
+                ],
+              )),
         ),
       ),
     );
   }
 
-  Future<List<dynamic>> getData(dynamic query) async {
+  Future<List<dynamic>> getData(dynamic query, String type) async {
     final response = await http
         .get(Uri.parse('https://ytmusic-tau.vercel.app/?search=$query'));
     var data = jsonDecode(response.body.toString());
     if (response.statusCode == 200) {
-      return data;
+      return data[type];
     }
-    return data;
+    return data[type];
   }
 }
