@@ -6,49 +6,63 @@ class PlaylistShimmer extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Shimmer.fromColors(
-      baseColor: Colors.white70,
-      highlightColor: Colors.white,
-      child: Column(
-        children: [
-          Center(
-            child: Padding(
-              padding: const EdgeInsets.only(top: 25.0, bottom: 25.0),
-              child: ClipRRect(
-                borderRadius: BorderRadius.circular(10.0),
-                child: Container(
-                  width: 250,
-                  height: 250,
-                  decoration: const BoxDecoration(color: Colors.black38),
+    return Column(
+      children: [
+        Container(
+          padding: const EdgeInsets.only(bottom: 20.0),
+          decoration: const BoxDecoration(
+              color: Color(0XFFC4FC4C),
+              borderRadius: BorderRadius.only(
+                  bottomLeft: Radius.circular(40.0),
+                  bottomRight: Radius.circular(40.0))),
+          child: Column(children: [
+            Center(
+              child: Padding(
+                padding: const EdgeInsets.only(top: 25.0, bottom: 25.0),
+                child: ClipRRect(
+                  borderRadius: BorderRadius.circular(10.0),
+                  child: Shimmer.fromColors(
+                    baseColor: Colors.black54,
+                    highlightColor: Colors.white,
+                    child: Container(
+                      width: 250,
+                      height: 250,
+                      decoration: const BoxDecoration(color: Colors.black38),
+                    ),
+                  ),
                 ),
               ),
             ),
-          ),
-          Center(
-            child: Container(
-              height: 20,
-              width: 250,
-              decoration: BoxDecoration(
-                  color: Colors.black38,
-                  borderRadius: BorderRadius.circular(10.0)),
+            Center(
+              child: Container(
+                height: 20,
+                width: 250,
+                decoration: BoxDecoration(
+                    color: Colors.black45,
+                    borderRadius: BorderRadius.circular(10.0)),
+              ),
             ),
-          ),
-          const SizedBox(
-            height: 15,
-          ),
-          Center(
-            child: Container(
-              height: 30,
-              width: 350,
-              decoration: const BoxDecoration(
-                  color: Colors.black38,
-                  borderRadius: BorderRadius.all(Radius.circular(10.0))),
+            const SizedBox(
+              height: 15,
             ),
-          ),
-          const SizedBox(
-            height: 30,
-          ),
-          Expanded(
+            Center(
+              child: Container(
+                height: 30,
+                width: 350,
+                decoration: const BoxDecoration(
+                    color: Colors.black38,
+                    borderRadius: BorderRadius.all(Radius.circular(10.0))),
+              ),
+            ),
+            const SizedBox(
+              height: 10,
+            )
+          ]),
+        ),
+        Expanded(
+          child: Shimmer.fromColors(
+            baseColor: Colors.white70,
+            highlightColor: Colors.white,
             child: ListView.builder(
                 scrollDirection: Axis.vertical,
                 itemCount: 10,
@@ -75,9 +89,9 @@ class PlaylistShimmer extends StatelessWidget {
                     ),
                   );
                 }),
-          )
-        ],
-      ),
+          ),
+        )
+      ],
     );
   }
 }
